backend/auth.py
import os
import secrets
from datetime import datetime, timedelta
from typing import Optional, Dict, Any
from jose import JWTError, jwt
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer

from backend.db import get_user_by_email, get_user_by_id

_env_key = os.getenv("HYDRAWATCH_SECRET_KEY")
if _env_key:
    SECRET_KEY = _env_key
else:
    SECRET_KEY = secrets.token_urlsafe(64)
    print("  ⚠ HYDRAWATCH_SECRET_KEY not set — generated ephemeral key.")
    print("    Sessions will NOT survive server restarts. Set the env var for production.")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 120 # 2 hours

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="api/auth/login")

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

async def get_current_user(token: str = Depends(oauth2_scheme)) -> Dict[str, Any]:
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_id: int = payload.get("sub")
        if user_id is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception

    user = get_user_by_id(int(user_id))
    if user is None:
        raise credentials_exception
    if user["status"] != "active":
        raise HTTPException(status_code=403, detail="Inactive user account")
    return user

async def get_current_active_admin(current_user: Dict[str, Any] = Depends(get_current_user)):
    if current_user["role"] != "admin":
        raise HTTPException(status_code=403, detail="Not enough permissions")
    return current_user

if __name__ == "__main__":
    # Bootstrap script for the first admin
    import sys
    from backend.db import create_user
    
    if len(sys.argv) != 3:
        print("Usage: python backend/auth.py <email> <password>")
        sys.exit(1)
        
    email = sys.argv[1]
    password = sys.argv[2]
    try:
        uid = create_user(email, get_password_hash(password), role="admin")
        print(f"✅ Admin user created with ID: {uid}")
    except Exception as e:
        print(f"❌ Failed to create admin: {e}")
