requirements.txt
wntr==1.2.0
torch==2.2.0
torch-geometric==2.5.0
xgboost==2.0.3
shap==0.44.1
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
websockets==12.0
openpyxl==3.1.2
matplotlib==3.8.3
onnx==1.15.0
onnxruntime==1.17.0
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
setuptools<70.0.0
